map Javascript's undefined 1:1 to new JSUndefined
original implementation had some issues in my opinion:
1. the mapping wasn't two-way, JS undefined was mapped to Python
None
2. there was a JSObjectUndefined wrapper available on Python side,
used to create Javascript undefined (for cases when it was explicitly
needed)

The problem with this approach is that it is not consistent.
JSObjectUndefined crossing the boundary into JS-land and then going
back to Python land does not maintain its identity.

Also it seems to be quite heavy-weight to maintain a full dynamically
allocated wrapper to represent constant JS undefined.

This new implementation introduces a new Python object which behaves like
None, but is distinct (internally it is implemented exactly like None in
CPython). This new object can represent JS undefined and work both ways.

Also we introduce JSNull shim into STPyV8 for new code not to hard-code
that JSNull is identical to Python's None.