CJSObjectFunction no longer holds m_self
this is a step towards flattening JSObject hierarchy as
mentioned here https://github.com/area1/stpyv8/issues/8#issuecomment-606702978

Q: So where is m_self when creating a wrapper for bound function?
A: We let JS engine create it and deal with wrapper for the bound function.

That means for given v8_fn and v8_self, we call `v8_fn.bind(v8_self)`
and this gives us v8_bound_fn to work with.
Our wrapper might be later asked to be applied to another "this",
we are lucky here because we can ask V8 to give us original unbound
function via v8_bound_fn->GetBoundFunction() and use that when needed.