introduce hospitals and hospitalize v8_errors there
Hospital is a place where we put V8 objects which have attached some
external data which need some cleanup. We call them "patients" which need
some care.

Motivation for this is to properly deal with raw Python objects
attached to JSException thrown around. The original code would leak them
if JSException object was not later translated for some reason.