stitch original PyV8 history
https://github.com/area1/stpyv8 forked PyV8 but dropped the history.
I prefer to have complete history of the project.

This commit takes the old PyV8 history up to [1],
then applies this delta commit to get to [2]
and then replays whole history up to [3].

Note that commits between [2] and [3] have new SHAs in this branch
because we replayed them on top of this commit.
I don't think it is a big issue. And I prefer it to merge commits.

[1] https://github.com/flier/pyv8/commit/d37f85ddc1c34a3e3c5e5fff272c2d9c975e147e
[2] https://github.com/darwin/stpyv8/commit/b54a0c403fa9ac2853f99384cd0e0b22d1d265d8
[3] https://github.com/darwin/stpyv8/commit/f4c216e02e90868952acc4a90004eae12d270d40