[API BREAK] we rely on pybind to extract error messages from python
1. after deeper reading of pybind code I realized py::error_already_set
has normalized values so we don't have to call PyErr_NormalizeException

2. also I realized that pybind does prepare a string error message in
the exception object which is almost what we want. So I take their
work and adapt it a bit.

This could break old code which relied on previous error message
extraction code. For example see the failing test which tested tuple
error value. Also errors relying on "message" or "args" attributes
could produce different messages.

I think it is a good step forward. Because we get rid of the hairy
code in extractExceptionMessage and rely on pybind for future.