PyErr_Occurred test is not needed in CJSEngine::ExecuteScript
I believe this test is not needed because if there is a python
error set, the python side should detect it upon returning back python.
Throwing py::error_already_set is going to be caught on pybind
boundary and the same python error would be restored back before
returning to python.

The above assumes the code is run as a call from Python side into C++
through pybind. If it is possible (or will be possible) to call this
in the context of C++ code with callstack not originating from Python
then checking and throwing python errors makes no sense. E.g. imagine
calling ExecuteScript from some C++ code to run pure javascript with no
python involvement at all.

Also we don't have any similar checks for PyErr_Occurred in our whole
naga codebase so that was my initial motivation to remove it.
It simply feels wrong.