ext: fix lock check in JSContext.__exit__
I believe this was a bug because testing hasattr(JSLocker, 'lock')
makes no sense to me. There should have been hasattr(self, 'lock')
instead.

It turns out that this was harmless in typical case because "del self"
on next line causes CLocker destructor to be called which properly
leaves/releases the lock. It just happens a bit later.

This bug would manifest itself in case when someone else were to hold
onto the lock after __exit__ for some reason.